backend/requirements/base.txt
Django==5.0.4
strawberry-graphql[django]==0.235.1
psycopg2-binary==2.9.9
django-environ==0.11.2
django-cors-headers==4.3.1
